[Tab Group] Fix undo logic for stable tab group IDs
Currently undoing with tab groups of size 1 is broken. Why?

1) The destination group isn't updated to remove the tab group ID.
2) The check in TabListMediator for isTabInTabGroup doesn't work for
   groups of size 1 if we need to add a new card when undoing a group
   action 1-by-1. Instead we should add a card when undoing and a
   tab or group of size 1 is not in the current set of tab cards.
3) Undo logic in TabGroupModelFilter wasn't consistently picking good
   candidate group IDs for mergeListOfTabsToGroup. Fix that.
4) mergeTabsToGroup was hard to read. Rearrange the method so it is
   more readable when adding the logic to include the destination
   tab in the undo logic.

Bug: b/41496693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5323801
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265476}

+++ chrome/android/features/tab_ui/java/src/org/chromium/chrome/browser/tasks/tab_management/UndoGroupSnackbarController.java
@@ -185,6 +185,8 @@ public class UndoGroupSnackbarController implements SnackbarManager.SnackbarCont
     public void onDismissNoAction(Object actionData) {
         // Delete the original tab group titles of the merging tabs once the merge is committed.
         for (TabUndoInfo info : (List<TabUndoInfo>) actionData) {
+            if (info.tab.getRootId() == info.tabOriginalRootId) continue;
+
             TabGroupTitleUtils.deleteTabGroupTitle(info.tabOriginalRootId);
         }
     }
